Remove the EnableMultipleConnections toggle
This is a PIR followup item to #incident-2025-10-23-snowflake-oauth-pdc-security-issue

We were respecting per-query arguments consistently when storing and
fetching database connections, leading to race conditions that may allow
a privilege escalations within a tenant.

This PR removes the distinction between a 'default' cache key and a key
with args, so we always take the same path when making a lookup key.

Also, as instances settings are one to one with a Connector, we stash
them on connector instances when creating them, so we don't need to fish
inside the default connection to create connections with query
arguments.

I wanted to rip out the default connection concept all together, but
this would have made the PR even bigger.

Also, I'm not sure what else we need to do to 'deprecate' this feature,
docs etc.